+++ microservice-core/include/MethodNotAllowedError.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "HttpError.hpp"
+
+class MethodNotAllowedError : public HttpError
+{
+public:
+    explicit MethodNotAllowedError(const std::string &msg = "Method not allowed")
+        : HttpError(405, msg) {}
+};